add media-sound/sndio and media-sound/aucatctl
adapted from vayerx overlay and xmw overlay respectively.

sndio was updated to 1.5.0 and the init script was changed to use openrc-run

aucatctl took some patches to make it build. first to make it compile
against libbsd (so we have strlcpy etc.), and then to fix some bad
format strings in the actual code.

TODO: investigate sndioctl (in the sndio git repo, mixer branch) as an
      alternative to aucatctl