ddrescueview: what ChatGPT got wrong
It got impressingly much right! It even knew it was on sourceforge, but
not how to get it. It did not get the license right. And I don't think
it actually ever looked inside the tarball, because it didn't know the
filenames in there (the filenames it did indeed know about was because I
told it about those files)

Some also just down to the ddrescueview being kinda non-standard...